+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * **Mod Support:**
   - Add support for optional menu and intermission sounds.
   - Add support for ambient sound definitions by SNDINFO lumps.
+  - Add support for optional chaingun sound, via `DSCHGUN`.
+
 
 * **Quality of Life:**
   - Player view no longer abruptly jolts when running across very shallow floor height changes